Moved pytest RBD fixtures to conftest.py file
